requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic
python-dotenv
pytest
httpx
jinja2
passlib[bcrypt]
python-jose
python-multipart
